07.Matzip-Agent-Python/pyproject.toml
[project]
name = "07-matzip-agent-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "langchain>=1.0.5",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.4",
    "langchain-openai>=1.0.2",
    "langchain-text-splitters>=1.0.0",
    "langgraph>=1.0.3",
    "langsmith>=0.4.42",
    "lxml>=6.0.2",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
]
